Implemented decorated collection proxy
When decorating collections, returning an array or decorators
for each element is not enough. Often the original collection
is richer then an array, e.g. has pagination methods.
This commit adds DecoratedEnumerableProxy that is returned
when decorating a collection of models, and which proxies
missing methods to the original wrapped collection.