Don't call ::scoped on an object if it's already a collection.
Before this commit, Draper::ModelSupport::ClassMethods#decorate calls
::scoped on the class that it was called upon, regardless of whether or
not it was already a collection. This has the property of removing any
extensions to the collection that may have already been applied, e.g. by
a pagination extension. Check for the presence of the :each method to
check if this is already a collection and decorate it directly if so.