Bugfix: do not confuse Kernel#Array
On 1.9.2 (and maybe earlier) Kernel#Array will choke on a decorator,
because it just invokes `to_ary` and `to_a` without using `respond_to?`
first, like it does in 1.9.3-p125.

As a work-around, make sure to not create a delegation method in
`method_missing` when the model will not respond to it anyway. The
decorator object will then behave like `Kernel#Array` expects it to.

See #115.