Adding cucumber stuff for integration testing.
I've embedded a little Rails app so that we can write cuke tests and
make sure that stuff doesn't break. Woo!

This was generated by making a rails engine, then copying just the
app over.

Thanks @nashby for fixing my stupid bug with sqlite3; it had to be
added as a development dependency. :heart: