Allow decorations of other decorators
Keep the "feature" of ignoring double-applications of the same
decorator. Figuring out a good behavior from the `initialize` method
where the model is already applied deeper in the chain was tricky, but I
hope a warning is enough.

I felt that raising an error would be too severe since there could be
valid cases for doing this, but on the other hand, it's probably a
mistake. If any valid cases show up I bet this will be reported as a
problem so we can re-evaluate this solution.