Remove Decorator#decorate
It's not clear what the expected behaviour should be here. Should
it just return self, and update options if passed? Should it create
a new decorator, in line with Decoratable#decorate?

Just ditching it and letting it fall through to the model seems to
be a better option than trying to work out what should happen when
you attempt to autodecorate an already-decorated object.