remove check for parent_class_name from decorator template
The only way I see to trigger the else case in this clause is to explicitly set the --parent option to something blank. I do not see why you would want to do this but I might be missing something here.

If this is on purpose it lacks a spec and the naming strategy is inconsistent in that it does not add the Decorator to the class_name.