Also initialize ViewContext for rails runner
Initialize Draper's ViewContext when using the runner command in a
similar way to using the console command. Without this code run by the
rails runner command (e.g. when processing commands from a message
queue) will not have a properly initialized ViewContext, and
consequently the helpers will not be initialized.