Document the fix for view context leaking in specs. (#847)
## Description
This pull request adds documentation for fixing the view context leakage between specs. I wanted to start with a documentation fix and go from there. If we later determine we need to clear the view context for all spec types by default, we can remove this documentation. See the referenced issue for the complete history of the issue.

## Testing
N/A

## To-Dos
None

## References
* [Should automatically clear view_context before/after each spec](https://github.com/drapergem/draper/issues/814)