Update BuildStrategy to be compatible with Ruby 2.7
When creating an ActionController::TestRequest, the params differ
based on the gem version of ActionPack.

Currently the check asserts the names of the params and expects
them to be [:req, :controller_class]. However on Ruby 2.7 this changes
to [:rest, :args].

Rather than check for the names of the params, we can make this more
robust by checking if there are any params. As we know that old versions
of TestRequest didn't require any params to create an instance.

The previous workaround was introduced in:
https://github.com/drapergem/draper/pull/759