Fix decoration spec
Draper::CollectionDecorator#{is_a?,kind_of?} triggers
decorating the object. When stubbing a method call
on a CollectionDecorator, RSpec itself will call #is_a?
on the object. See:

https://github.com/rspec/rspec-mocks/blob/ea4d8ea4532480500462cc7a633d3408656a4dcd/lib/rspec/mocks/proxy.rb#L38

That's why the object will always be decorated when
stubbing a method on it.

We can circumvent this issue by directly checking
the definition of @decorated_collection.