Enforce to disable `ActiveSupport::ExecutionContext.nestable`
`ActiveSupport::ExecutionContext.nestable` would be true when
`ActiveSupport::TestCase` is loaded.
https://github.com/rails/rails/blob/b47cc51b989e09e8a00616bc336c8da7347a90d5/activesupport/lib/active_support/railtie.rb#L68

We load `action_controller/test_case` in Rails console, so this
would be true.

But, it looks like ``ActiveSupport::ExecutionContext.nestable`
only consider to enable on test environment. That cause unexpected
behavior. To avoid an error, enforce to disable it.

Fixes #948.